api/requirements.txt
flask==2.0.1
flask-cors==3.0.10
google-cloud-firestore==2.3.4
gunicorn==20.1.0 